requirements.txt
-e .
aws-cdk.aws-ecs==1.122.0
aws-cdk.aws-ecs-patterns==1.122.0
aws-cdk.aws-route53==1.122.0
flask
